+++ packages/database/src/mikro-orm/unit-of-work.ts
@@ -13,6 +13,7 @@ import { MikroOrmDocumentRepository } from './repositories/DocumentRepository.js
 import { MikroOrmUserPromptRepository } from './repositories/UserPromptRepository.js';
 import { MikroOrmTaskRepository } from './repositories/TaskRepository.js';
 import { MikroOrmClaudeMdRepository } from './repositories/ClaudeMdRepository.js';
+import { MikroOrmCodeSnippetRepository } from './repositories/CodeSnippetRepository.js';
 import { MikroOrmDailyStatsRepository } from './repositories/DailyStatsRepository.js';
 import { MikroOrmTechnologyUsageRepository } from './repositories/TechnologyUsageRepository.js';
 import { MikroOrmAchievementRepository } from './repositories/AchievementRepository.js';
@@ -28,6 +29,7 @@ export class MikroOrmUnitOfWork implements IUnitOfWork {
   public userPrompts: MikroOrmUserPromptRepository;
   public taskQueue: MikroOrmTaskRepository;
   public claudemd: MikroOrmClaudeMdRepository;
+  public codeSnippets: MikroOrmCodeSnippetRepository;
   public dailyStats: MikroOrmDailyStatsRepository;
   public technologyUsage: MikroOrmTechnologyUsageRepository;
   public achievements: MikroOrmAchievementRepository;
@@ -42,6 +44,7 @@ export class MikroOrmUnitOfWork implements IUnitOfWork {
     this.userPrompts = new MikroOrmUserPromptRepository(em);
     this.taskQueue = new MikroOrmTaskRepository(em);
     this.claudemd = new MikroOrmClaudeMdRepository(em);
+    this.codeSnippets = new MikroOrmCodeSnippetRepository(em);
     this.dailyStats = new MikroOrmDailyStatsRepository(em);
     this.technologyUsage = new MikroOrmTechnologyUsageRepository(em);
     this.achievements = new MikroOrmAchievementRepository(em);
@@ -93,6 +96,7 @@ export class MikroOrmUnitOfWork implements IUnitOfWork {
     this.userPrompts = new MikroOrmUserPromptRepository(em);
     this.taskQueue = new MikroOrmTaskRepository(em);
     this.claudemd = new MikroOrmClaudeMdRepository(em);
+    this.codeSnippets = new MikroOrmCodeSnippetRepository(em);
     this.dailyStats = new MikroOrmDailyStatsRepository(em);
     this.technologyUsage = new MikroOrmTechnologyUsageRepository(em);
     this.achievements = new MikroOrmAchievementRepository(em);